@@ -1,6 +1,6 @@
 {
   "name": "medic-app",
-  "version": "1.0.1",
+  "version": "1.3.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
@@ -11,58 +11,62 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "7.2.9",
-    "@angular/common": "7.2.9",
-    "@angular/compiler": "7.2.9",
-    "@angular/core": "7.2.9",
-    "@angular/forms": "7.2.9",
-    "@angular/http": "7.2.9",
-    "@angular/platform-browser": "7.2.9",
-    "@angular/platform-browser-dynamic": "7.2.9",
-    "@angular/router": "7.2.9",
-    "core-js": "2.5.4",
-    "rxjs": "6.4.0",
-    "tslib": "1.9.0",
-    "zone.js": "0.8.29",
-    "@ngrx/store": "7.2.0",
-    "ngx-malihu-scrollbar": "7.0.0",
-    "ng2-dragula": "2.1.1",
-    "ng2-charts": "1.6.0",
-    "chart.js": "2.7.3",
-    "d3": "5.7.0",
-    "@swimlane/ngx-charts": "10.0.0",
-    "echarts": "4.2.0-rc.2",
-    "echarts-gl": "1.1.1",
-    "ngx-echarts": "4.1.0",
     "@agm/core": "1.0.0-beta.5",
-    "leaflet": "1.4.0",
-    "@asymmetrik/ngx-leaflet": "5.0.1",
-    "ng-fullcalendar": "1.7.1",
-    "fullcalendar": "3.6.1",
-    "animate.css": "3.7.0",
-    "web-animations-js": "2.3.1",
-    "@types/jquery": "3.3.29"
+    "@angular/animations": "9.0.0-next.0",
+    "@angular/common": "9.0.0-next.0",
+    "@angular/compiler": "9.0.0-next.0",
+    "@angular/core": "9.0.0-next.0",
+    "@angular/fire": "^5.2.1",
+    "@angular/forms": "9.0.0-next.0",
+    "@angular/platform-browser": "9.0.0-next.0",
+    "@angular/platform-browser-dynamic": "9.0.0-next.0",
+    "@angular/router": "9.0.0-next.0",
+    "@asymmetrik/ngx-leaflet": "5.0.2",
+    "@fullcalendar/angular": "4.1.1",
+    "@fullcalendar/core": "4.1.0",
+    "@fullcalendar/daygrid": "4.1.0",
+    "@ngrx/store": "8.2.0",
+    "@swimlane/ngx-charts": "11.1.0",
+    "@types/jquery": "3.3.29",
+    "angularfire2": "^5.2.1",
+    "animate.css": "3.7.1",
+    "chart.js": "2.8.0",
+    "core-js": "3.1.3",
+    "d3": "5.9.2",
+    "echarts": "4.2.1",
+    "echarts-gl": "1.1.1",
+    "firebase": "^6.3.5",
+    "leaflet": "1.5.1",
+    "ng2-charts": "2.2.4",
+    "ng2-dragula": "2.1.1",
+    "ngx-echarts": "4.1.1",
+    "ngx-malihu-scrollbar": "7.0.0",
+    "rxjs": "6.5.2",
+    "tslib": "1.9.3",
+    "web-animations-js": "2.3.2",
+    "zone.js": "0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.12.3",
-    "@angular/cli": "7.3.2",
-    "@angular/compiler-cli": "7.2.5",
-    "@angular/language-service": "7.2.5",
-    "@types/jasmine": "3.3.8",
-    "@types/jasminewd2": "2.0.3",
-    "@types/leaflet": "1.4.0",
-    "@types/node": "10.12.18",
-    "codelyzer": "4.5.0",
-    "jasmine-core": "3.3.0",
+    "@angular-devkit/build-angular": "0.802.0",
+    "@angular/cli": "8.2.0",
+    "@angular/compiler-cli": "9.0.0-next.0",
+    "@angular/language-service": "9.0.0-next.0",
+    "@types/jasmine": "3.3.13",
+    "@types/jasminewd2": "2.0.6",
+    "@types/leaflet": "1.4.4",
+    "@types/node": "12.0.4",
+    "codelyzer": "5.0.1",
+    "jasmine-core": "3.4.0",
     "jasmine-spec-reporter": "4.2.1",
-    "karma": "4.0.0",
+    "karma": "4.1.0",
     "karma-chrome-launcher": "2.2.0",
-    "karma-coverage-istanbul-reporter": "2.0.1",
+    "karma-coverage-istanbul-reporter": "2.0.5",
     "karma-jasmine": "2.0.1",
-    "karma-jasmine-html-reporter": "1.4.0",
-    "protractor": "5.4.0",
-    "ts-node": "8.0.1",
-    "tslint": "5.12.1",
-    "typescript": "3.2.4"
+    "karma-jasmine-html-reporter": "1.4.2",
+    "node-sass": "4.12.0",
+    "protractor": "5.4.2",
+    "ts-node": "8.2.0",
+    "tslint": "5.17.0",
+    "typescript": "3.5.3"
   }
 }
