@@ -17,12 +17,12 @@ export class Page404Component extends BasePageComponent implements OnInit, OnDes
     super(store, httpSv);
 
     this.pageData = {
-      title: 'Page not found!',
+      title: 'En Proceso...!',
       loaded: true,
       breadcrumbs: [
         {
-          title: 'Dashboard',
-          route: 'default-dashboard'
+          title: 'patients',
+          route: 'patients'
         },
         {
           title: 'Error!'
